requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
python-slugify==6.1.1
requests==2.27.1
text-unidecode==1.3
urllib3==1.26.9

pip~=22.0.4
wheel~=0.37.1
setuptools~=61.0.0